RUM-14814: Split ViewEventDiff into diff logic and type conversions
