+++ features/dd-sdk-android-rum/src/main/kotlin/com/datadog/android/rum/internal/model/ViewEventDiff.kt
@@ -4,8 +4,6 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-@file:Suppress("TooManyFunctions")
-
 package com.datadog.android.rum.internal.model
 
 import com.datadog.android.internal.utils.computeDiffIfChanged
@@ -249,183 +247,6 @@ private fun diffAccessibility(
             rtlEnabled = diffEquals(ViewEvent.Accessibility::rtlEnabled)
         )
     }
-}
-
-// endregion
-
-// region Type conversions
-
-private fun ViewEvent.Usr.toRum() = ViewUpdateEvent.Usr(
-    id = id,
-    name = name,
-    email = email,
-    anonymousId = anonymousId,
-    additionalProperties = additionalProperties
-)
-
-private fun ViewEvent.Account.toRum() = ViewUpdateEvent.Account(
-    id = id,
-    name = name,
-    additionalProperties = additionalProperties
-)
-
-private fun ViewEvent.Connectivity.toRum() = ViewUpdateEvent.Connectivity(
-    status = status.toRum(),
-    interfaces = interfaces?.map { it.toRum() },
-    effectiveType = effectiveType?.toRum(),
-    cellular = cellular?.let { ViewUpdateEvent.Cellular(it.technology, it.carrierName) }
-)
-
-private fun ViewEvent.Display.toRum() = ViewUpdateEvent.Display(
-    scroll = scroll?.let {
-        ViewUpdateEvent.Scroll(it.maxDepth, it.maxDepthScrollTop, it.maxScrollHeight, it.maxScrollHeightTime)
-    },
-    viewport = viewport?.let { ViewUpdateEvent.Viewport(it.width, it.height) }
-)
-
-private fun ViewEvent.Os.toRum() = ViewUpdateEvent.Os(
-    name = name,
-    version = version,
-    build = build,
-    versionMajor = versionMajor
-)
-
-private fun ViewEvent.Device.toRum() = ViewUpdateEvent.Device(
-    type = type?.toRum(),
-    name = name,
-    model = model,
-    brand = brand,
-    architecture = architecture,
-    locale = locale,
-    locales = locales,
-    timeZone = timeZone,
-    batteryLevel = batteryLevel,
-    powerSavingMode = powerSavingMode,
-    brightnessLevel = brightnessLevel,
-    logicalCpuCount = logicalCpuCount,
-    totalRam = totalRam,
-    isLowRam = isLowRam
-)
-
-private fun ViewEvent.Container.toRum() = ViewUpdateEvent.Container(
-    view = ViewUpdateEvent.ContainerView(view.id),
-    source = source.toRumSource()
-)
-
-private fun ViewEvent.FlutterBuildTime.toRum() = ViewUpdateEvent.FlutterBuildTime(
-    min = min,
-    max = max,
-    average = average,
-    metricMax = metricMax
-)
-
-private fun ViewEvent.PerformanceCls.toRum() = ViewUpdateEvent.Cls(
-    score = score,
-    timestamp = timestamp,
-    targetSelector = targetSelector,
-    previousRect = previousRect?.let { ViewUpdateEvent.PreviousRect(it.x, it.y, it.width, it.height) },
-    currentRect = currentRect?.let { ViewUpdateEvent.PreviousRect(it.x, it.y, it.width, it.height) }
-)
-
-private fun ViewEvent.Inp.toRum() = ViewUpdateEvent.Inp(
-    duration = duration,
-    timestamp = timestamp,
-    targetSelector = targetSelector
-)
-
-private fun ViewEvent.Lcp.toRum() = ViewUpdateEvent.Lcp(
-    timestamp = timestamp,
-    targetSelector = targetSelector,
-    resourceUrl = resourceUrl
-)
-
-// endregion
-
-// region Enum conversions
-
-private fun ViewEvent.ViewEventSource.toRumSource() = when (this) {
-    ViewEvent.ViewEventSource.ANDROID -> ViewUpdateEvent.ViewUpdateEventSource.ANDROID
-    ViewEvent.ViewEventSource.IOS -> ViewUpdateEvent.ViewUpdateEventSource.IOS
-    ViewEvent.ViewEventSource.BROWSER -> ViewUpdateEvent.ViewUpdateEventSource.BROWSER
-    ViewEvent.ViewEventSource.FLUTTER -> ViewUpdateEvent.ViewUpdateEventSource.FLUTTER
-    ViewEvent.ViewEventSource.REACT_NATIVE -> ViewUpdateEvent.ViewUpdateEventSource.REACT_NATIVE
-    ViewEvent.ViewEventSource.ROKU -> ViewUpdateEvent.ViewUpdateEventSource.ROKU
-    ViewEvent.ViewEventSource.UNITY -> ViewUpdateEvent.ViewUpdateEventSource.UNITY
-    ViewEvent.ViewEventSource.KOTLIN_MULTIPLATFORM -> ViewUpdateEvent.ViewUpdateEventSource.KOTLIN_MULTIPLATFORM
-}
-
-private fun ViewEvent.ViewEventSessionType.toRum() = when (this) {
-    ViewEvent.ViewEventSessionType.USER -> ViewUpdateEvent.ViewUpdateEventSessionType.USER
-    ViewEvent.ViewEventSessionType.SYNTHETICS -> ViewUpdateEvent.ViewUpdateEventSessionType.SYNTHETICS
-    ViewEvent.ViewEventSessionType.CI_TEST -> ViewUpdateEvent.ViewUpdateEventSessionType.CI_TEST
-}
-
-private fun ViewEvent.LoadingType.toRum() = when (this) {
-    ViewEvent.LoadingType.INITIAL_LOAD -> ViewUpdateEvent.LoadingType.INITIAL_LOAD
-    ViewEvent.LoadingType.ROUTE_CHANGE -> ViewUpdateEvent.LoadingType.ROUTE_CHANGE
-    ViewEvent.LoadingType.ACTIVITY_DISPLAY -> ViewUpdateEvent.LoadingType.ACTIVITY_DISPLAY
-    ViewEvent.LoadingType.ACTIVITY_REDISPLAY -> ViewUpdateEvent.LoadingType.ACTIVITY_REDISPLAY
-    ViewEvent.LoadingType.FRAGMENT_DISPLAY -> ViewUpdateEvent.LoadingType.FRAGMENT_DISPLAY
-    ViewEvent.LoadingType.FRAGMENT_REDISPLAY -> ViewUpdateEvent.LoadingType.FRAGMENT_REDISPLAY
-    ViewEvent.LoadingType.VIEW_CONTROLLER_DISPLAY -> ViewUpdateEvent.LoadingType.VIEW_CONTROLLER_DISPLAY
-    ViewEvent.LoadingType.VIEW_CONTROLLER_REDISPLAY -> ViewUpdateEvent.LoadingType.VIEW_CONTROLLER_REDISPLAY
-}
-
-private fun ViewEvent.ConnectivityStatus.toRum() = when (this) {
-    ViewEvent.ConnectivityStatus.CONNECTED -> ViewUpdateEvent.Status.CONNECTED
-    ViewEvent.ConnectivityStatus.NOT_CONNECTED -> ViewUpdateEvent.Status.NOT_CONNECTED
-    ViewEvent.ConnectivityStatus.MAYBE -> ViewUpdateEvent.Status.MAYBE
-}
-
-private fun ViewEvent.Interface.toRum() = when (this) {
-    ViewEvent.Interface.BLUETOOTH -> ViewUpdateEvent.Interface.BLUETOOTH
-    ViewEvent.Interface.CELLULAR -> ViewUpdateEvent.Interface.CELLULAR
-    ViewEvent.Interface.ETHERNET -> ViewUpdateEvent.Interface.ETHERNET
-    ViewEvent.Interface.WIFI -> ViewUpdateEvent.Interface.WIFI
-    ViewEvent.Interface.WIMAX -> ViewUpdateEvent.Interface.WIMAX
-    ViewEvent.Interface.MIXED -> ViewUpdateEvent.Interface.MIXED
-    ViewEvent.Interface.OTHER -> ViewUpdateEvent.Interface.OTHER
-    ViewEvent.Interface.UNKNOWN -> ViewUpdateEvent.Interface.UNKNOWN
-    ViewEvent.Interface.NONE -> ViewUpdateEvent.Interface.NONE
-}
-
-private fun ViewEvent.EffectiveType.toRum() = when (this) {
-    ViewEvent.EffectiveType.SLOW_2G -> ViewUpdateEvent.EffectiveType.SLOW_2G
-    ViewEvent.EffectiveType.`2G` -> ViewUpdateEvent.EffectiveType.`2G`
-    ViewEvent.EffectiveType.`3G` -> ViewUpdateEvent.EffectiveType.`3G`
-    ViewEvent.EffectiveType.`4G` -> ViewUpdateEvent.EffectiveType.`4G`
-}
-
-private fun ViewEvent.DeviceType.toRum() = when (this) {
-    ViewEvent.DeviceType.MOBILE -> ViewUpdateEvent.DeviceType.MOBILE
-    ViewEvent.DeviceType.DESKTOP -> ViewUpdateEvent.DeviceType.DESKTOP
-    ViewEvent.DeviceType.TABLET -> ViewUpdateEvent.DeviceType.TABLET
-    ViewEvent.DeviceType.TV -> ViewUpdateEvent.DeviceType.TV
-    ViewEvent.DeviceType.GAMING_CONSOLE -> ViewUpdateEvent.DeviceType.GAMING_CONSOLE
-    ViewEvent.DeviceType.BOT -> ViewUpdateEvent.DeviceType.BOT
-    ViewEvent.DeviceType.OTHER -> ViewUpdateEvent.DeviceType.OTHER
-}
-
-private fun ViewEvent.ReplayLevel.toRum() = when (this) {
-    ViewEvent.ReplayLevel.ALLOW -> ViewUpdateEvent.ReplayLevel.ALLOW
-    ViewEvent.ReplayLevel.MASK -> ViewUpdateEvent.ReplayLevel.MASK
-    ViewEvent.ReplayLevel.MASK_USER_INPUT -> ViewUpdateEvent.ReplayLevel.MASK_USER_INPUT
-}
-
-private fun ViewEvent.Plan.toRum() = when (this) {
-    ViewEvent.Plan.PLAN_1 -> ViewUpdateEvent.Plan.PLAN_1
-    ViewEvent.Plan.PLAN_2 -> ViewUpdateEvent.Plan.PLAN_2
-}
-
-private fun ViewEvent.SessionPrecondition.toRum() = when (this) {
-    ViewEvent.SessionPrecondition.USER_APP_LAUNCH -> ViewUpdateEvent.SessionPrecondition.USER_APP_LAUNCH
-    ViewEvent.SessionPrecondition.INACTIVITY_TIMEOUT -> ViewUpdateEvent.SessionPrecondition.INACTIVITY_TIMEOUT
-    ViewEvent.SessionPrecondition.MAX_DURATION -> ViewUpdateEvent.SessionPrecondition.MAX_DURATION
-    ViewEvent.SessionPrecondition.BACKGROUND_LAUNCH -> ViewUpdateEvent.SessionPrecondition.BACKGROUND_LAUNCH
-    ViewEvent.SessionPrecondition.PREWARM -> ViewUpdateEvent.SessionPrecondition.PREWARM
-    ViewEvent.SessionPrecondition.FROM_NON_INTERACTIVE_SESSION ->
-        ViewUpdateEvent.SessionPrecondition.FROM_NON_INTERACTIVE_SESSION
-    ViewEvent.SessionPrecondition.EXPLICIT_STOP -> ViewUpdateEvent.SessionPrecondition.EXPLICIT_STOP
 }
 
 // endregion
